fix: Separator 로직 hooks -> contextAPI로 이동
